pyproject.toml
[tool.poetry]
name = "split-mmd"
version = "0.1.0"
description = ""
authors = ["rosetteer-kim <rosetteer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
watchdog = "^5.0.3"
pymongo = "^4.10.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
